CSCFAIRMETA-718: Qvain navbar * When running Qvain in separate site, show different navbar content * Make header components more general so Etsin and Qvain can share theme * Add Qvain logo * Resize Qvain and Etsin logo images to be same size * Fix logout not working properly on narrow screen
