+++ etsin_finder/frontend/js/components/header/etsinLogo.jsx
@@ -1,0 +1,20 @@
+{
+  /**
+   * This file is part of the Etsin service
+   *
+   * Copyright 2017-2018 Ministry of Education and Culture, Finland
+   *
+   *
+   * @author    CSC - IT Center for Science Ltd., Espoo Finland <servicedesk@csc.fi>
+   * @license   MIT
+   */
+}
+
+import React from 'react'
+
+import Logo from '../general/header/logo'
+import EtsinLogoImage from '../../../static/images/Etsin_300px.png'
+
+const EtsinLogo = () => <Logo image={EtsinLogoImage} alt="Fairdata Etsin logo" />
+
+export default EtsinLogo
