+++ etsin_finder/frontend/js/routes/index.jsx
@@ -10,11 +10,11 @@
    */
 }
 import loadable from '@loadable/component'
-// eslint-disable-next-line
+
 export const Home = loadable(() => import('../components/frontpage'))
 
 export const Dataset = loadable(() => import('../components/dataset' /* webpackPrefetch: true */))
-// eslint-disable-next-line
+
 export const Search = loadable(() => import('../components/search' /* webpackPrefetch: true */))
 
 export const Qvain = loadable(() => import('../components/qvain/main' /* webpackPrefetch: true */))
