+++ etsin_finder/frontend/locale/finnish.js
@@ -343,6 +343,11 @@ const finnish = {
     },
   },
   qvain: {
+    nav: {
+      home: 'Koti',
+      createDataset: 'Luo aineisto',
+      editDataset: 'Muokkaa aineistoa',
+    },
     general: {
       brief: 'Tutkimusaineiston metatietotyökalu',
       looseActors: {
